requirements.txt
django
psycopg2
jinja2
django-markdownx

